requirements.txt
imap-tools==1.0.0
pdfminer.six==20221105
python-dotenv==0.21.0
